requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
aiosqlite==0.19.0
bcrypt==4.1.2
python-multipart==0.0.6
